requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
cffi==1.14.6
click==8.0.1
cryptg==0.2.post4
essential-generators==1.0
fastapi==0.68.0
h11==0.12.0
Jinja2==3.0.1
MarkupSafe==2.0.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.19.0
rsa==4.7.2
starlette==0.14.2
Telethon==1.23.0
typing-extensions==3.10.0.0
uvicorn==0.15.0
